move river tests to folder
